Option to move Fonts out of HBIOS area to ROM Bank 3
The FONTS_INLINE equate controls whether the fonts are included in the HBIOS area as per existing behaviour or whether they are placed in the currently unused ROM bank 3.
In the case where we use bank 3 for font storage, the HBIOS is built with the font data located at 0x8000 onwards. The build process splits this off for relocation to the correct image position.  Unlzsa, the TMS and VGA driver have been updated to access the font data from either location, whether compressed or not.

Moving the font data to bank 3 results in significant savings in HBIOS space.

If you are happy with this approach, I will continue developing and roll changes into other display drivers and do some hardware testing.

Further on, as mentioned, I would like to look see if i can get the decompressor to decompress one byte at a time and move the CCP cache to bank 3 as well.

Regards Phil